Refactor Publish modal
